Updated edit form color
