+++ frontend/lib/routes/session/SessionScreen.dart
@@ -11,6 +11,7 @@ import 'package:frontend/routes/member/EditMemberForm.dart';
 import 'package:frontend/routes/session/DisplayGeneralInformation.dart';
 import 'package:frontend/routes/session/DisplayTickets.dart';
 import 'package:frontend/routes/session/EditSessionForm.dart';
+import 'package:frontend/routes/session/SessionBanner.dart';
 import 'package:frontend/routes/session/SessionInformationBox.dart';
 import 'package:frontend/routes/session/SessionsNotifier.dart';
 import 'package:frontend/services/authService.dart';
@@ -79,7 +80,7 @@ class _SessionScreen extends State<SessionScreen>
                   DisplayGeneralInformation(
                     session: widget.session,
                   ),
-                  DisplaySpeaker(
+                  DisplaySpeakers(
                     session: widget.session,
                   ),
                   DisplayTickets(session: widget.session),
@@ -90,273 +91,3 @@ class _SessionScreen extends State<SessionScreen>
     });
   }
 }
-
-class SessionBanner extends StatefulWidget {
-  final Session session;
-
-  const SessionBanner({Key? key, required this.session}) : super(key: key);
-
-  @override
-  _SessionBannerState createState() => _SessionBannerState();
-}
-
-class _SessionBannerState extends State<SessionBanner> {
-  SessionService _sessionService = SessionService();
-  editMember() {
-    return FutureBuilder(
-        future: Provider.of<AuthService>(context).role,
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            Role r = snapshot.data as Role;
-
-            if (r == Role.ADMIN || r == Role.COORDINATOR) {
-              return Positioned(
-                  bottom: 15,
-                  right: 15,
-                  child: GestureDetector(
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) =>
-                                EditSessionForm(session: widget.session)),
-                      );
-                    },
-                    child: Container(
-                      height: 40,
-                      width: 40,
-                      decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        border: Border.all(
-                          width: 2,
-                          color: Theme.of(context).scaffoldBackgroundColor,
-                        ),
-                        color: Colors.indigo[200],
-                      ),
-                      child: Icon(
-                        Icons.edit,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ));
-            } else
-              return Container(width: 0);
-          } else
-            return Container(width: 0);
-        });
-  }
-
-  void _deleteSessionDialog(context, id) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return BlurryDialog(
-            'Warning', 'Are you sure you want to delete session?', () {
-          _deleteSession(context, id);
-        });
-      },
-    );
-  }
-
-  void _deleteSession(context, id) async {
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('Deleting')),
-    );
-
-    Session? s = await _sessionService.deleteSession(id);
-    if (s != null) {
-      SessionsNotifier notifier =
-          Provider.of<SessionsNotifier>(context, listen: false);
-      notifier.remove(s);
-
-      ScaffoldMessenger.of(context).hideCurrentSnackBar();
-
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text('Done'),
-          duration: Duration(seconds: 2),
-        ),
-      );
-    } else {
-      ScaffoldMessenger.of(context).hideCurrentSnackBar();
-
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('An error occured.')),
-      );
-    }
-  }
-
-  Future<void> _editSessionModal(context, id) async {
-    Future<Session> sessionFuture = _sessionService.getSession(id);
-    Session session = await sessionFuture;
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) {
-        return FractionallySizedBox(
-            heightFactor: 0.7,
-            child: Container(
-              child: EditSessionForm(session: session),
-            ));
-      },
-    );
-    setState(() {});
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    int event = Provider.of<EventNotifier>(context).event.id;
-    return LayoutBuilder(builder: (context, constraints) {
-      bool small = constraints.maxWidth < App.SIZE;
-      return Container(
-        width: constraints.maxWidth,
-        decoration: BoxDecoration(
-          image: DecorationImage(
-            image: AssetImage("assets/banner_background.png"),
-            fit: BoxFit.cover,
-          ),
-        ),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: <Widget>[
-            SizedBox(height: 30),
-            // Stack(children: [
-            //   Container(
-            //     width: 210,
-            //     height: 210,
-            //     decoration: BoxDecoration(
-            //       shape: BoxShape.circle,
-            //       border: Border.all(color: Colors.white30),
-            //     ),
-            //     padding: const EdgeInsets.all(5),
-            //     // child: Hero(
-            //     //   tag: widget.member.id + event.toString(),
-            //     //   child: ClipOval(
-            //     //     child: (widget.member.image == '')
-            //     //         ? Image.asset("assets/noImage.png")
-            //     //         : Image.network(widget.member.image!),
-            //     //   ),
-            //     // ),
-            //   ),
-            //   // editMember(),
-            // ]),
-            SizedBox(height: 20),
-            Text(widget.session.title,
-                style: TextStyle(
-                    fontSize: 25,
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold)),
-            SizedBox(height: 10),
-            IconButton(
-                onPressed: () {
-                  _editSessionModal(context, widget.session.id);
-                },
-                icon: Icon(Icons.edit),
-                color: const Color(0xff5c7ff2)),
-            FutureBuilder(
-                future: Provider.of<AuthService>(context).role,
-                builder: (context, snapshot) {
-                  if (snapshot.hasData) {
-                    Role r = snapshot.data as Role;
-
-                    if (r == Role.ADMIN || r == Role.COORDINATOR) {
-                      return IconButton(
-                          onPressed: () =>
-                              _deleteSessionDialog(context, widget.session.id),
-                          icon: Icon(Icons.delete),
-                          color: Colors.red);
-                    } else {
-                      return Container();
-                    }
-                  } else {
-                    return Container();
-                  }
-                }),
-            SizedBox(height: 20),
-          ],
-        ),
-      );
-    });
-  }
-}
-
-// class DisplayTickets extends StatefulWidget {
-//   final Session session;
-//   final bool small;
-//   const DisplayTickets({Key? key, required this.session, required this.small})
-//       : super(key: key);
-
-//   @override
-//   _DisplayTicketsState createState() => _DisplayTicketsState();
-// }
-
-// class _DisplayTicketsState extends State<DisplayTickets> {
-//   @override
-//   void initState() {
-//     super.initState();
-//   }
-
-//   @override
-//   Widget build(BuildContext conext) {
-//     return Scaffold(
-//       backgroundColor: Color.fromRGBO(186, 196, 242, 0.1),
-//       body: ListView(
-//         padding: EdgeInsets.symmetric(horizontal: 32),
-//         physics: BouncingScrollPhysics(),
-//         children: [
-//           (widget.session.tickets!.max == 0)
-//               ? Text("There") //NoTicketsAvailable(widget.session)
-//               : Text("Here"),
-//           // SessionInformationBox(session: widget.session, type: "description"),
-//           // InformationBox(title: "Phones", contact: cont, type: "phone"),
-//           // InformationBox(
-//           //     title: "Socials",
-//           //     contact: cont,
-//           //     type: "social"), //SizedBox(height: 24,),
-//         ],
-//       ),
-//       // floatingActionButton: _isEditable(cont),
-//     );
-//   }
-
-//   Widget NoTicketsAvailable(Session session) {
-//     return Container(
-//       margin: EdgeInsets.fromLTRB(0, 20, 0, 0),
-//       padding: EdgeInsets.fromLTRB(17, 15, 17, 15),
-//       decoration: BoxDecoration(
-//           color: Colors.white,
-//           borderRadius: BorderRadius.circular(5),
-//           boxShadow: [
-//             BoxShadow(
-//                 blurRadius: 7.0,
-//                 color: Colors.grey.withOpacity(0.3),
-//                 offset: new Offset(0, 3),
-//                 spreadRadius: 4.0),
-//           ]),
-//       child: Stack(children: [
-//         Column(
-//           crossAxisAlignment: CrossAxisAlignment.start,
-//           children: [
-//             Text("Info on tickets",
-//                 textAlign: TextAlign.left,
-//                 style: TextStyle(
-//                     fontSize: 22,
-//                     color: Colors.black,
-//                     fontWeight: FontWeight.bold)),
-//             Divider(),
-//             Row(
-//               mainAxisAlignment: MainAxisAlignment.spaceBetween,
-//               children: [
-//                 Text(
-//                   "No tickets available for this session.",
-//                   textAlign: TextAlign.left,
-//                   style: TextStyle(fontSize: 16, color: Colors.black),
-//                 ),
-//               ],
-//             )
-//           ],
-//         ),
-//       ]),
-//     );
-//   }
-// }
